% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/daily_hl.R
\name{daily_hl}
\alias{daily_hl}
\title{Daily Tide Values}
\usage{
daily_hl(df, level, daytime)
}
\arguments{
\item{level}{character of column containing gauge data.}

\item{daytime}{character of column containing daytime values from lubridate.}

\item{data}{frame containg tide data and day time.}
}
\description{
Pulls out higher high, high, low, lower low tides from daily gauge values.
Expects a single day as input or grouped output (i.e. from dplyr::group_by)
}
\examples{
file<-system.file("extdata","EDC_WaterLevel_Test.csv",package="tider")
tides<-read.csv(file,stringsAsFactor=F)
tides$DateTime <- lubridate::parse_date_time(paste(tides$Date,tides$Time),
                                 "\%m/\%d/\%y \%H:\%M:\%S")
daily_hl(tides$Depth_NAVD88m,tides$DateTime)
}

