% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tider_byday.R
\name{tider_byday}
\alias{tider_byday}
\title{Return a data frame of tide metrics per day}
\usage{
tider_byday(df, level, daytime)
}
\arguments{
\item{level}{character of column containing gauge data.}

\item{daytime}{character of column containing daytime values from lubridate.}

\item{data}{frame containg tide data and day time.}
}
\description{
Return a data frame of tide metrics per day
}
\examples{
file<-system.file("extdata","TideFile2days.csv",package="tider")
tides<-read.csv(file,stringsAsFactor=F)
tides$DateTime <- lubridate::parse_date_time(paste(tides$Date,tides$Time),
                                 "\%m/\%d/\%y \%H:\%M:\%S")
tider_byday(tides, "Depth_NAVD88m", "DateTime")
}

