% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tider_byday_dep.R
\name{tider_byday_dep}
\alias{tider_byday_dep}
\title{Return a data frame of tide metrics per day}
\usage{
tider_byday_dep(df, level, daytime)
}
\arguments{
\item{level}{character of column containing gauge data.}

\item{daytime}{character of column containing daytime values from lubridate.}

\item{data}{frame containg tide data and day time.}
}
\description{
Return a data frame of tide metrics per day
}

