% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tider_byday.R
\name{tider_hl}
\alias{tider_hl}
\title{Daily Tide Values}
\usage{
tider_hl(level, daytime, metric = c("hh", "h", "l", "ll"))
}
\arguments{
\item{level}{numeric vector containing gauge data.}

\item{daytime}{POSIXct vector of daytime for a single day}
}
\description{
Pulls out higher high, high, low, lower low tides from daily gauge values.
Expects a single day as input or grouped output (i.e. from dplyr::group_by)
}
\examples{
file<-system.file("extdata","EDC_WaterLevel_Test.csv",package="tider")
tides<-read.csv(file,stringsAsFactor=F)
tides$DateTime <- lubridate::parse_date_time(paste(tides$Date,tides$Time),
                                 "\%m/\%d/\%y \%H:\%M:\%S")
daily_hl(tides, "Depth_NAVD88m", "DateTime")
}

