% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tider_correct.R
\name{tider_correct}
\alias{tider_correct}
\title{Calculate corrected spring, neap, etc.}
\usage{
tider_correct(mhw_acc, s2_ctrl, mn_acc, type = c("spring", "neap"))
}
\arguments{
\item{mhw_acc}{A numeric of the Mean High Water of the nearest NOAA Control Station}

\item{s2_ctrl}{A numeric of the Prinicipal solor semidiurnal constituent from the nearest NOAA
control station.}

\item{mn_acc}{A numeric of the Mean range from nearest NOAA control station}

\item{type}{the corrected value to return}
}
\value{
returns a numeric of the corrected spring or neap
}
\description{
Need better description here
}

